feat(app-platform): Web Installation Flow
App developers will have two options for installation paths. First, a
purely async path that relies on an installation webhook. This is useful
when you don't need any information about Users and just want to push
data into Sentry.

The second is when you do need information about the User or to
correlate data between Sentry and the App in some way.

This change adds the latter (just in Developer Settings for now). When
the User clicks the install button, it creates the record on our side
and redirects the User to the App. It includes data about the
installation and everything the App needs to do the token exchange.